SNA/requirements.txt
asgiref==3.3.1
certifi==2020.11.8
Django==3.0.8
django-admin-interface==0.16.3
django-better-admin-arrayfield==1.4.2
django-colorfield==0.4.2
django-filebrowser-no-grappelli==3.8.0
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-tinymce==3.3.0
django-tinymce4-lite==1.8.0
jsmin==2.2.2
Markdown==3.3.3
Pillow==8.0.1
python-dotenv==0.19.0
pytz==2020.4
six==1.14.0
sqlparse==0.4.1
wincertstore==0.2
